refactor: move components to their own folders
