+++ apps/web/app/post/[pubky]/[postId]/components/_ReplyForm.tsx
@@ -11,7 +11,7 @@ import { useClientContext } from '@/contexts';
 import Modal from '@/components/Modal';
 import { Utils } from '@social/utils-shared';
 import { IPost } from '@/types';
-import LinkPreviewer from '@/components/_LinkPreviewer';
+import LinkPreviewer from '@/components/LinkPreview';
 
 export default function ReplyForm({
   uri,
