+++ apps/web/components/AtiveFriends/index.tsx
@@ -5,7 +5,7 @@ import { useEffect, useState } from 'react';
 import { useClientContext } from '@/contexts';
 import { Utils } from '@social/utils-shared';
 import { IFriendsResponse } from '@/types';
-import Skeletons from './Skeletons';
+import Skeletons from '../Skeletons';
 
 export default function ActiveFriends() {
   const { pubky, listFollowers, listFollowing } = useClientContext();
