+++ apps/web/components/CreateQuickPost/index.tsx
@@ -10,9 +10,9 @@ import EmojiPicker, {
 import { useClientContext, useAlertContext } from '@/contexts';
 import Image from 'next/image';
 import { INewPost } from '@/types';
-import Modal from './Modal';
+import Modal from '../Modal';
 import { Utils } from '@social/utils-shared';
-import LinkPreviewer from './_LinkPreviewer';
+import LinkPreviewer from '../LinkPreview';
 
 export default function CreateQuickPost() {
   const { pubky, getProfile, createPost, setPosts, createTag } =
