+++ apps/web/components/Header/index.tsx
@@ -11,7 +11,7 @@ import {
   Button,
   Menu,
 } from '@social/ui-shared';
-import { Modal } from './Modal';
+import { Modal } from '../Modal';
 import { useClientContext } from '@/contexts';
 import { Utils } from '@social/utils-shared';
