+++ apps/web/components/Modal/_Repost.tsx
@@ -18,7 +18,7 @@ import { Modal as ModalComponent } from '.';
 import { Utils } from '@social/utils-shared';
 import { IPost } from '@/types';
 import Post from '../Post';
-import LinkPreviewer from '@/components/_LinkPreviewer';
+import LinkPreviewer from '@/components/LinkPreview';
 
 interface CreateRepostProps {
   showModalRepost: boolean;
