+++ apps/web/components/RemindBackup/index.tsx
@@ -3,7 +3,7 @@
 import { z } from 'zod';
 import { Button, Icon, Typography } from '@social/ui-shared';
 import { useEffect, useRef, useState } from 'react';
-import Modal from './Modal';
+import Modal from '../Modal';
 import { useClientContext } from '@/contexts';
 import { Utils } from '@social/utils-shared';
