+++ apps/web/components/Skeletons/index.tsx
@@ -1,11 +1,9 @@
 import DisplayText from './_DisplayText';
-import Post from './_Post';
 import ProfileCard from './_ProfileCard';
 import Simple from './_Simple';
 
 export const Skeletons = {
   DisplayText,
-  Post,
   ProfileCard,
   Simple,
 };
